Instantiate data loader only once
